+++ src/logstory/__init__.py
@@ -19,4 +19,4 @@ from importlib.metadata import PackageNotFoundError, version
 try:
   __version__ = version("logstory")
 except PackageNotFoundError:
-  __version__ = "1.2.2"
+  __version__ = "1.2.3"